+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Privacy level is not getting saved property [#2666](https://github.com/rokwire/illinois-app/issues/2666)
+
+## [4.2.37] - 2022-12-09
+### Added
+- Show "Cancelled" label for Appointments [#2692](https://github.com/rokwire/illinois-app/issues/2692).
+### Fixed
+- Fixed Teams Coach and Roaster list panels open full size image url[#2694](https://github.com/rokwire/illinois-app/issues/2694).
+- McKinley link and phone number handling[#2659](https://github.com/rokwire/illinois-app/issues/2659).
+### Changed
+- MTD Stops and Desinations map view updates [#2633](https://github.com/rokwire/illinois-app/issues/2633).
 
 ## [4.2.36] - 2022-12-08
 ### Changed